refactor(types): relocate misplaced helpers to single-responsibility modules
Two ownership mismatches from before af06485:

- BRIDGE_WAIT_SPAWNED_TIMEOUT_MS lived in godot-runner.ts but was
  exported only because runtime-tools.ts needed the default. Move to
  bridge-protocol.ts alongside the other shared bridge constants.

- handler-helpers.ts mixed two concerns with disjoint consumer sets.
  executeSceneOp (used by scene-tools + node-tools) stays in a renamed
  utils/headless-op.ts. The runtime-only helpers (parseBridgeJson,
  attachRuntimeWarnings, MAX_RUNTIME_ERROR_CONTEXT_LINES, ParseResult)
  inline into runtime-tools.ts -- their only consumer.

Pure mechanical move -- no behaviour change, no signature change.
Updates every import site, renames the unit test, and refreshes
CLAUDE.md / docs/architecture.md / docs/tool-authoring.md.
